requirements.txt
psycopg[binary]
cryptography
python-dotenv
yfinance
apscheduler
pandas
numpy
pytest
pytest-mock
akshare